Update font configuration
Blocking DejaVu fallbacks needs to happen at the system level with
latest fontconfig. I've also cleaned things up significantly while
debugging this and learning about fontconfig. Unfortunately, emoji no
longer render properly in rofi and dunst, which seems to be due to the
fontconfig update and not my configuration, but I'm not sure.