emacs: Fix desktop entries
I had several issues with these.

All of them appeared in menus, because I used `True` for `NoDisplay`
instead of `true`.

The Elisp snippets weren't evaluating correctly due to issues with
escaping in the standard for the `Exec` key, so I moved the
`emacsclient` call to a script. Furthermore, the functions I was calling
were not loaded when Emacs was launched, which meant they only worked
when Mu4e and Elpher were already loaded.

I misunderstood StartupWMClass, which meant all Emacs windows were
treated as Elpher by the desktop environment. Only the desired "parent"
needs that defined.

I used `desktop-file-utils` to discover the `True` capitalization error
and to `update-desktop-database` while experimenting.