monolith: Remove vfio-pci binding at boot
I've learned that libvirt handles unbinding device drivers and binding
vfio-pci on VM startup. None of this is actually necessary in my case.
It could be necessary if GPU reset bugs were an issue or if my host had
a graphical session as well, but that's not the case for me.