emacs: Add override for Doom project name
The directory variables method Projectile suggests
https://docs.projectile.mx/projectile/projects.html doesn't work here
because the workspace naming happens before the directory variables are
loaded. Furthermore, Doom calls the function even if the variable is
set, which should probably be fixed.

This is a hack and requires modification for every override, but it
works for me for now.