Add default bookmark location to tridactylrc
I had to use JavaScript via `js` to concatenate the path, and I'm
replacing slashes (solidus characters) in the title with the full-width
solidus variant from Unicode to get around Tridactyl interpreting them
as path separators.

This places bookmarks in the Mobile Bookmarks folder by default, which I
prefer because the default can't be changed from that folder on iOS.
This allows me to have a unified bookmark inbox for export and triage
into my notes.